+++ packages/issue-utils/CHANGELOG.md
@@ -1,5 +1,16 @@
 # Changelog
 
+## [2.0.0](https://github.com/googleapis/repo-automation-bots/compare/issue-utils-v1.0.0...issue-utils-v2.0.0) (2022-08-15)
+
+
+### ⚠ BREAKING CHANGES
+
+* **bot-config-utils:** use gcf-utils 14 (#4143)
+
+### deps
+
+* **bot-config-utils:** use gcf-utils 14 ([#4143](https://github.com/googleapis/repo-automation-bots/issues/4143)) ([2295665](https://github.com/googleapis/repo-automation-bots/commit/22956655ed839268725fa75f1bc11ee856e9e281))
+
 ## 1.0.0 (2022-07-08)